+++ gulpfile.js
@@ -1,30 +1,45 @@
 const gulp = require("gulp");
-const install = require('gulp-install');
-const shell = require('gulp-shell');
+const gutil = require("gulp-util");
+const tslint = require("gulp-tslint");
+const ts = require("gulp-typescript");
+const nodemon = require("nodemon");
+const path = require("path");
 
-// Tasks relating build
-gulp.task('build:server', shell.task('cd participant_server && npm run build'));
+const tsProject = ts.createProject("./tsconfig.json");
 
-
-gulp.task('build', gulp.series(
-    'build:server'
-));
-
-// Tasks relating dependencies
-
-gulp.task('dependencies:server', () => {
-    return gulp.src(['./participant_server/package.json']).pipe(install());
+gulp.task("watch", (done) => {
+    gulp.watch("src/**/*", gulp.series("lint", "build"));
 });
 
-gulp.task('dependencies', gulp.series(
-    'dependencies:server',
-));
+gulp.task("lint", () => {
+    return gulp.src(["src/**/*.ts"])
+        .pipe(tslint({
+            formatter: "prose",
+        }))
+        .pipe(tslint.report({
+            emitError: false,
+        }));
+});
 
-// Postinstall
-gulp.task('postinstall', gulp.series(
-    'dependencies',
-    'build',
-));
+gulp.task("build:server", () => {
+    return tsProject.src()
+        .pipe(tsProject())
+        .pipe(gulp.dest("build/"));
+});
 
-// Default task
-gulp.task('default', gulp.parallel('dependencies'));
+gulp.task("build", gulp.series("lint", "build:server"));
+
+gulp.task("serve", gulp.series("lint", "build:server", () => {
+    nodemon({
+        script: path.join(__dirname, "build/app.js"),
+        watch: ["build/"],
+        ignore: ["build/public", "./node_modules"],
+        env: {
+            "NODE_ENV": "dev",
+        },
+    }).on("start", function() {
+        gutil.log(gutil.colors.blue("Server started!"));
+    });
+}));
+
+gulp.task("default", gulp.parallel("serve", "watch"));
