--- server/package.json
@@ -1,53 +1,0 @@
-{
-  "name": "virtual-checkin-server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.js",
-  "scripts": {
-    "start": "gulp",
-    "serve": "cd src/ && ts-node app.ts",
-    "build": "gulp build"
-  },
-  "keywords": [],
-  "author": "Mohan Dodda",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/compression": "^1.7.0",
-    "@types/cors": "^2.8.6",
-    "@types/dotenv": "^8.2.0",
-    "@types/express": "^4.17.7",
-    "@types/express-graphql": "^0.9.0",
-    "@types/express-session": "^1.17.0",
-    "@types/mongoose": "^5.7.30",
-    "@types/morgan": "^1.9.1",
-    "@types/passport": "^1.0.4",
-    "@types/passport-oauth2": "^1.4.9",
-    "@types/request": "^2.48.5",
-    "gulp-tslint": "^8.1.4",
-    "nodemon": "^2.0.4",
-    "ts-node": "^8.10.1",
-    "tslint": "^6.1.2",
-    "typescript": "^3.9.7"
-  },
-  "dependencies": {
-    "apollo-server": "^2.17.0",
-    "apollo-server-express": "^2.17.0",
-    "body-parser": "^1.19.0",
-    "compression": "^1.7.4",
-    "connect-mongo": "^3.2.0",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "express-graphql": "^0.7.1",
-    "express-session": "^1.17.1",
-    "fs": "0.0.1-security",
-    "graphql": "^14.1.1",
-    "mongoose": "^5.9.22",
-    "morgan": "^1.10.0",
-    "node-fetch": "^2.6.1",
-    "passport": "^0.4.1",
-    "passport-oauth2": "^1.5.0",
-    "path": "^0.12.7",
-    "request": "^2.88.2"
-  }
-}
